+++ tests/conftest.py
@@ -1,6 +1,8 @@
 import os
+from unittest.mock import MagicMock
 
 import pytest
+from prefect.testing.utilities import prefect_test_harness
 
 from prefect_snowflake.credentials import SnowflakeCredentials
 
@@ -16,6 +18,26 @@ def _read_test_file(name: str) -> bytes:
     full_name = os.path.join(os.path.split(__file__)[0], "test_data", name)
     with open(full_name, "rb") as fd:
         return fd.read()
+
+
+@pytest.fixture(scope="session", autouse=True)
+def prefect_db():
+    """
+    Sets up test harness for temporary DB during test runs.
+    """
+    with prefect_test_harness():
+        yield
+
+
+@pytest.fixture(autouse=True)
+def reset_object_registry():
+    """
+    Ensures each test has a clean object registry.
+    """
+    from prefect.context import PrefectObjectRegistry
+
+    with PrefectObjectRegistry():
+        yield
 
 
 @pytest.fixture()
@@ -93,3 +115,10 @@ def private_malformed_credentials_params():
         "password": "letmein",
         "private_key": _read_test_file("test_cert_malformed_format.p8"),
     }
+
+
+@pytest.fixture
+def snowflake_connect_mock(monkeypatch):
+    connect_mock = MagicMock()
+    monkeypatch.setattr("snowflake.connector.connect", connect_mock)
+    return connect_mock